Move deleteAccountRequest to separate file
Pass headers object to deleteAccountRequest function.
