+++ app/setup-account-request/setup-account-request.js
@@ -1,6 +1,5 @@
-const { accessTokenAndResourcePath, consentAccountRequestId, removeAuthServerUserConsent } = require('../authorise');
-const { postAccountRequests, deleteAccountRequest } = require('./account-requests');
-const { session } = require('../session');
+const { accessTokenAndResourcePath } = require('../authorise');
+const { postAccountRequests } = require('./account-requests');
 
 const createRequest = async (resourcePath, accessToken, fapiFinancialId) => {
   const headers = { accessToken, fapiFinancialId };
@@ -23,42 +22,6 @@ const createRequest = async (resourcePath, accessToken, fapiFinancialId) => {
   throw error;
 };
 
-
-const deleteRequest = async (
-  sessionId,
-  authorisationServerId,
-  fapiFinancialId,
-  fapiInteractionId,
-) => {
-  const fail = () => {
-    const error = new Error('Bad Request');
-    error.status = 400;
-    throw error;
-  };
-  const username = await session.getUsername(sessionId);
-  const { accessToken, resourcePath } = await accessTokenAndResourcePath(authorisationServerId);
-  const keys = { username, authorisationServerId, scope: 'accounts' };
-  const accountRequestId = await consentAccountRequestId(keys);
-  if (!accountRequestId) return fail();
-  const responseHeaders = await deleteAccountRequest(
-    resourcePath,
-    accessToken,
-    fapiFinancialId,
-    accountRequestId,
-    fapiInteractionId,
-  );
-  if (responseHeaders) {
-    const interactionId = responseHeaders['x-fapi-interaction-id'];
-    if (fapiInteractionId === interactionId) {
-      await removeAuthServerUserConsent(keys);
-      return 204;
-    }
-    return fail();
-  }
-  return fail();
-};
-
-
 const setupAccountRequest = async (authorisationServerId, fapiFinancialId) => {
   const { accessToken, resourcePath } = await accessTokenAndResourcePath(authorisationServerId);
   const accountRequestId = await createRequest(
@@ -70,4 +33,3 @@ const setupAccountRequest = async (authorisationServerId, fapiFinancialId) => {
 };
 
 exports.setupAccountRequest = setupAccountRequest;
-exports.deleteRequest = deleteRequest;
